Updated tsconfig.json settings
The TypeScript configuration file has been updated. The "jsx" setting was changed from "react-jsx" to "preserve". An "incremental" property was added and set to true for faster subsequent builds. Also, ".next/types/**/*.ts" was included in the files to be compiled, and a minor formatting issue at the end of the file was fixed.